Hice un par de refactor y limpie codigo que ya no se usar
